Add call to domain inference
